remove rest of unsused consts + move used ones in constants folder and update imports
