+++ packages/veraswap-app/src/routes/index.lazy.tsx
@@ -4,14 +4,9 @@ import { useAccount, useReadContract, useSwitchChain, useWatchContractEvent, use
 import {
     getSwapAndHyperlaneSweepBridgeTransaction,
     getSwapExactInExecuteData,
-    MAX_UINT_160,
-    MAX_UINT_256,
-    MAX_UINT_48,
-    PERMIT2_ADDRESS,
     getHyperlaneMessageIdFromReceipt,
     getSwapAndSuperchainBridgeTransaction,
     getSuperchainMessageIdFromReceipt,
-    SUPERCHAIN_TOKEN_BRIDGE,
     getTransaction,
     TransactionParams,
     HypERC20Token,
@@ -23,7 +18,13 @@ import { useAtom, useAtomValue } from "jotai";
 import { useEffect } from "react";
 import { ProcessId } from "@owlprotocol/contracts-hyperlane/artifacts/IMailbox";
 import { RelayedMessage } from "@owlprotocol/veraswap-sdk/artifacts/IL2ToL2CrossDomainMessenger";
-import { UNISWAP_CONTRACTS } from "@owlprotocol/veraswap-sdk/constants";
+import {
+    MAX_UINT_160,
+    MAX_UINT_256,
+    MAX_UINT_48,
+    PERMIT2_ADDRESS,
+    UNISWAP_CONTRACTS,
+} from "@owlprotocol/veraswap-sdk/constants";
 import {
     hyperlaneGasPaymentAtom,
     quoteInAtom,
