+++ packages/veraswap-sdk/src/chains/supersim.ts
@@ -1,5 +1,8 @@
 import { Chain, localhost } from "viem/chains";
 import { createPublicClient, http } from "viem";
+// TODO: delete?
+export const SUPERCHAIN_SWEEP_ADDRESS = "0x7eF899a107a9a98002E23910838731562A3e8dC4";
+export const SUPERCHAIN_TOKEN_BRIDGE = "0x4200000000000000000000000000000000000028";
 
 /*** Local Supersim Chains ***/
 export const opChainL1Port = 8547;
