+++ packages/veraswap-sdk/src/constants/divvi.ts
@@ -1,0 +1,1 @@
+export const DIVVI_BASE_REGISTRY = "0xba9655677f4e42dd289f5b7888170bc0c7da8cdc";
