+++ packages/veraswap-sdk/src/constants/index.ts
@@ -1,3 +1,4 @@
+export * from "./divvi.js";
 export * from "./hyperlane.js";
 export * from "./kernel.js";
 export * from "./tokens.js";
