+++ packages/veraswap-sdk/src/constants/uniswap.ts
@@ -10,6 +10,8 @@ import { UniversalRouterApprovedReentrant as UniversalRouter } from "../artifact
 import { opChainA, opChainB, opChainL1 } from "../chains/supersim.js";
 import { arbitrum, arbitrumSepolia, base, bsc, mainnet, sepolia } from "viem/chains";
 
+export const V4_SWAP = 0x10;
+
 /*** Uniswap Constants ***/
 export const PERMIT2_ADDRESS = "0x000000000022D473030F116dDEE9F6B43aC78BA3";
 /**
