+++ packages/veraswap-sdk/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./hooks/index.js";
 export * from "./query/index.js";
-export * from "./constants.js";
 export * from "./swap/index.js";
 export * from "./types/index.js";
 export * from "./uniswap/index.js";
