+++ packages/veraswap-sdk/src/scripts/deployTestTokens.ts
@@ -23,14 +23,20 @@ import { IERC20 } from "@owlprotocol/contracts-hyperlane";
 import { CurrencyAmount, Price, Token } from "@uniswap/sdk-core";
 import { Pool, V4PositionPlanner, priceToClosestTick, Position } from "@uniswap/v4-sdk";
 import { MockERC20 } from "../artifacts/MockERC20.js";
-import { MAX_UINT_160, MAX_UINT_256, MAX_UINT_48, PERMIT2_ADDRESS, testHyperlaneRegistry } from "../constants.js";
 import { getChainNameAndMailbox } from "../utils/getChainNameAndMailbox.js";
 import { IAllowanceTransfer } from "../artifacts/IAllowanceTransfer.js";
 import { IPositionManager } from "../artifacts/IPositionManager.js";
 import { IMulticall_v4 } from "../artifacts/IMulticall_v4.js";
 import { PoolKeyAbi } from "../types/PoolKey.js";
 import { IStateView } from "../artifacts/IStateView.js";
-import { UNISWAP_CONTRACTS } from "../constants/uniswap.js";
+import {
+    PERMIT2_ADDRESS,
+    UNISWAP_CONTRACTS,
+    MAX_UINT_160,
+    MAX_UINT_256,
+    MAX_UINT_48,
+    testHyperlaneRegistry,
+} from "../constants/index.js";
 
 // TODO: just have this in the sdk, and import later in app
 const fetchGithubRegistryData = async () => {
