+++ packages/veraswap-sdk/src/swap/getSwapAndHyperlaneSweepBridgeTransaction.ts
@@ -4,7 +4,7 @@ import { getV4SwapCommandParams } from "./getV4SwapCommandParams.js";
 import { CommandType, RoutePlanner } from "../uniswap/routerCommands.js";
 import { PoolKey } from "../types/PoolKey.js";
 import { IUniversalRouter } from "../artifacts/IUniversalRouter.js";
-import { HYPERLANE_ROUTER_SWEEP_ADDRESS } from "../constants.js";
+import { HYPERLANE_ROUTER_SWEEP_ADDRESS } from "../constants/index.js";
 
 /**
  * getSwapAndHyperlaneSweepBridgeTransaction generates a transaction for the Uniswap Router to swap tokens and bridge them to another chain using Hyperlane
