+++ packages/veraswap-sdk/src/swap/getSwapAndSuperchainBridgeTransaction.ts
@@ -4,7 +4,7 @@ import { getV4SwapCommandParams } from "./getV4SwapCommandParams.js";
 import { CommandType, RoutePlanner } from "../uniswap/routerCommands.js";
 import { PoolKey } from "../types/PoolKey.js";
 import { IUniversalRouter } from "../artifacts/IUniversalRouter.js";
-import { SUPERCHAIN_SWEEP_ADDRESS } from "../constants.js";
+import { SUPERCHAIN_SWEEP_ADDRESS } from "../chains/index.js";
 
 /**
  * getSwapAndSuperchainBridgeTransaction generates a transaction for the Uniswap Router to swap tokens and bridge them to another chain using Superchain Interop
