+++ packages/veraswap-sdk/src/swap/getTransaction.ts
@@ -31,7 +31,7 @@ export interface TransactionSwapBridgeOptions {
 }
 
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface TransactionBridgeSwapOptions { }
+export interface TransactionBridgeSwapOptions {}
 
 export type TransactionParams =
     | (TransactionTypeSwap & TransactionSwapOptions)
